Move ApacheHttpGetResponseEntityContentContinuatorIT to the right package
