+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/client/MantaClient.java
@@ -221,7 +221,28 @@ public class MantaClient implements AutoCloseable {
      *
      * @param config The configuration context that provides all of the configuration values
      * @param connectionFactoryConfigurator pre-configured objects for use with a MantaConnectionFactory (or null)
+     * @param metricConfiguration the metrics registry and configuration, or null to prepare one from the general config
+     */
+    public MantaClient(final ConfigContext config,
+                       final MantaConnectionFactoryConfigurator connectionFactoryConfigurator,
+                       final MantaClientMetricConfiguration metricConfiguration) {
+        this(config, connectionFactoryConfigurator, null, metricConfiguration);
+    }
+
+    /**
+     * Creates a new instance of the Manta client based on user-provided connection objects. This allows for a higher
+     * degree of customization at the cost of more involvement from the consumer.
+     *
+     * Users opting into advanced configuration (i.e. not passing {@code null} as the second parameter)
+     * should be comfortable with the internals of {@link CloseableHttpClient} and accept that we can only make a
+     * best effort to support all possible use-cases. For example, users may pass in a builder which is wired to a
+     * {@link org.apache.http.impl.conn.BasicHttpClientConnectionManager} and effectively make the client
+     * single-threaded by eliminating the connection pool. Bug or feature? You decide!
+     *
+     * @param config The configuration context that provides all of the configuration values
+     * @param connectionFactoryConfigurator pre-configured objects for use with a MantaConnectionFactory (or null)
      * @param httpHelper helper object for executing http requests (or null to build one ourselves)
+     * @param metricConfiguration the metrics registry and configuration, or null to prepare one from the general config
      */
     MantaClient(final ConfigContext config,
                 final MantaConnectionFactoryConfigurator connectionFactoryConfigurator,
